requirements.txt
fastapi
pydantic
requests
python-dotenv
pyjwt
cryptography
uvicorn
pandas
python-multipart
pynamodb